Cleanup of the sdba doc - move some functions to processing - move of api pages
