+++ docs/api.rst
@@ -2,6 +2,8 @@
 API
 ===
 
+The API of the statistical downscaling and bias adjustment module (sdba) is documented :ref:`on this page <SDBA's user API>`. The API of the ``cfchecks``, ``datachecks``, ``missing`` and ``dataflags`` modules are in :ref:`Health checks`. Finally, the API of the translating tools is on the :ref:`Internationalization` page.
+
 
 Indicators
 ==========
@@ -44,7 +46,6 @@ Indicator tools
    :member-order: bysource
    :show-inheritance:
 
-
 Unit handling module
 ====================
 
@@ -52,13 +53,6 @@ Unit handling module
    :members:
    :undoc-members:
    :show-inheritance:
-
-Statistical Downscaling and Bias Adjustment
-===========================================
-
-.. toctree::
-
-     sdba_api
 
 
 Other utilities
