--- docs/sdba_api.rst
@@ -1,28 +1,0 @@
-.. automodule:: xclim.sdba
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-.. _bias-adjustment-algos:
-
-Available methods
------------------
-
-.. automodule:: xclim.sdba.adjustment
-   :members:
-   :special-members:
-   :show-inheritance:
-
-
-Utilities
----------
-
-.. automodule:: xclim.sdba.utils
-   :members: equally_spaced_nodes
-   :undoc-members:
-   :show-inheritance:
-
-.. automodule:: xclim.sdba.processing
-   :members: adapt_freq, jitter_under_thresh
-   :undoc-members:
-   :show-inheritance:
