+++ xclim/sdba/__init__.py
@@ -36,16 +36,16 @@ Modular approach
 This module adopts a modular approach instead of implementing published and named methods directly.
 A generic bias adjustment process is laid out as follows:
 
-- preprocessing on `ref`, `hist` and `sim` (using methods in `xclim.sdba.processing` or `xclim.sdba.detrending`)
-- creating and training the adjustment object `Adj = Adjustment.train(obs, sim, **kwargs)` (from `xclim.sdba.adjustment`)
-- adjustment `scen = Adj.adjust(sim, **kwargs)`
-- post-processing on `scen` (for example: re-trending)
+- preprocessing on ``ref``, ``hist`` and ``sim`` (using methods in :py:mod:`xclim.sdba.processing` or :py:mod:`xclim.sdba.detrending`)
+- creating and training the adjustment object ``Adj = Adjustment.train(obs, sim, **kwargs)`` (from :py:mod:`xclim.sdba.adjustment`)
+- adjustment ``scen = Adj.adjust(sim, **kwargs)``
+- post-processing on ``scen`` (for example: re-trending)
 
 The train-adjust approach allows to inspect the trained adjustment object. The training information is stored in
 the underlying `Adj.ds` dataset and always has a `af` variable with the adjustment factors. Its layout and the
-other available variables vary between the different algorithm, refer to :ref:`bias-adjustment-algos`.
+other available variables vary between the different algorithm, refer to :ref:`Adjustment methods`.
 
-Parameters needed by the training and the adjustment are saved to the `Adj.ds` dataset as a  `adj_params` attribute.
+Parameters needed by the training and the adjustment are saved to the ``Adj.ds`` dataset as a  `adj_params` attribute.
 Other parameters, those only needed by the adjustment are passed in the `adjust` call and written to the history attribute
 in the output scenario dataarray.
 
@@ -55,7 +55,7 @@ Grouping
 For basic time period grouping (months, day of year, season), passing a string to the methods needing it is sufficient.
 Most methods acting on grouped data also accept a `window` int argument to pad the groups with data from adjacent ones.
 Units of `window` are the sampling frequency of the main grouping dimension (usually `time`). For more complex grouping,
-one can pass a :py:class:`xclim.sdba.base.Grouper` directly.
+one can pass an instance of :py:class:`xclim.sdba.base.Grouper` directly.
 
 Notes for developers
 ====================
@@ -66,12 +66,12 @@ this reason, the module is often divided in two parts : the (decorated) compute 
 and the user-facing functions or objects in corresponding public file (ex: ``adjustment.py``). See the `sdba-advanced`
 notebook for more info on the reasons for this move.
 
-Other restrictions : `map_blocks` will remove any "auxiliary" coordinates before calling the wrapped function and will add them back on exit.
+Other restrictions : ``map_blocks`` will remove any "auxiliary" coordinates before calling the wrapped function and will add them back on exit.
 """
 from . import detrending, processing, utils
 from .adjustment import *
-from .base import (
-    Grouper,
+from .base import Grouper
+from .processing import (
     construct_moving_yearly_window,
     stack_variables,
     unpack_moving_yearly_window,
