+++ xclim/sdba/nbutils.py
@@ -1,4 +1,7 @@
-"""Numba-accelerated utils."""
+"""
+Numba-accelerated utilities
+---------------------------
+"""
 import numpy as np
 from numba import boolean, float32, float64, guvectorize, njit
 from xarray import DataArray
